Add service to manipulate zfs crds
